warning fixes: mostly size_t vs. specialized API type and other type conversion. added player_id_t typedef and INVALID_PLAYER, use that instead of -1. also added sanity checks to cpu.cpp to ensure ARCH_* is correct (see http://www.wildfiregames.com/forum/index.php?showtopic=13327&hl=) and further predefined macros to arch.h just to be sure.
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@8079 3db68df2-c116-0410-a063-a993310a9797

+++ source/collada/StdSkeletons.cpp
@@ -200,7 +200,7 @@ void Skeleton::LoadSkeletonDataFromXml(const char* xmlData, size_t xmlLength, st
 	try
 	{
 		xmlSetGenericErrorFunc(&xmlErrors, &errorHandler);
-		doc = xmlParseMemory(xmlData, xmlLength);
+		doc = xmlParseMemory(xmlData, (int)xmlLength);
 		if (doc)
 		{
 			xmlNode* root = xmlDocGetRootElement(doc);
